tests: move compilation and prediction tests to test_untrained_model.py
